Inception V3 model for Forest Fire Detection
The concept of the Inception v3 is to focus on consumption of lesser computational power which is achieved by altering the already existing inception architectures. When compared to other architectures like VGGNet, Inception models are proved to be more computationally efficient in terms of economical cost and number of parameters. In the earlier versions of the inception model, adaption of the network for different cases results in problem because of the uncertainty of the new network’s efficiency. Whereas in InceptionV3 model many techniques like factorized convolutions, regularization, dimension reduction, and parallelized computations are incorporated for optimizing the network and to loosen the constraints for adaption.