Inception Resnet V2 model
Inception-ResNet v1 and v2 are the two versions of this hybrid. Despite the fact that their working concepts are similar, Inception-ResNet v2 is more precise.
This allows the model to take advantage of the benefits of both the Inception series of networks, which allows filters to run several kernel sizes on the same network, and the residual network, which allows neural networks to use skip connections, or shortcuts, to skip over some layers. More than a million photos from the ImageNet collection were used to train it.